# server/src/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Oya Browser API",
    "description": "REST API for controlling browsers connected to the Oya Browser server. Browsers connect via WebSocket; AI tools and clients use this REST API (or MCP) to send commands.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://oyabrowser.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3100",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/browsers/{browserId}/command": {
      "post": {
        "summary": "Send a command to a browser",
        "description": "Execute a browser action (analyze, navigate, click, type, press_key, screenshot, scroll, wait, list_tabs, open_tab, switch_tab, close_tab, read_page).",
        "operationId": "sendCommand",
        "parameters": [
          { "$ref": "#/components/parameters/browserId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "analyze",
                      "navigate",
                      "click",
                      "type",
                      "press_key",
                      "screenshot",
                      "scroll",
                      "wait",
                      "read_page",
                      "list_tabs",
                      "open_tab",
                      "switch_tab",
                      "close_tab"
                    ],
                    "description": "The browser action to execute. Each action uses only specific params (see examples)."
                  },
                  "params": {
                    "type": "object",
                    "description": "Action-specific parameters. Only include params relevant to the chosen action.",
                    "properties": {
                      "url": { "type": "string", "description": "URL (for navigate, open_tab)" },
                      "selector": { "type": "string", "description": "CSS selector, e.g. [data-ac-id=\"3\"] (for click, type, wait, read_page)" },
                      "text": { "type": "string", "description": "Text to type (for type)" },
                      "key": { "type": "string", "description": "Key to press (for press_key): Enter, Escape, Tab, Backspace, ArrowDown, ArrowUp, etc." },
                      "direction": { "type": "string", "enum": ["up", "down"], "description": "Scroll direction (for scroll)" },
                      "amount": { "type": "number", "description": "Pixels to scroll (for scroll, default 500)" },
                      "timeout": { "type": "number", "description": "Timeout in ms (for wait, default 10000)" },
                      "tab_id": { "type": "number", "description": "Tab ID (for switch_tab, close_tab)" },
                      "limit": { "type": "number", "description": "Max elements (for read_page, default 50)" }
                    }
                  }
                },
                "examples": [
                  { "action": "navigate", "params": { "url": "https://google.com" } },
                  { "action": "analyze" },
                  { "action": "click", "params": { "selector": "[data-ac-id=\"3\"]" } },
                  { "action": "type", "params": { "selector": "[data-ac-id=\"9\"]", "text": "hello world" } },
                  { "action": "press_key", "params": { "key": "Enter" } },
                  { "action": "scroll", "params": { "direction": "down", "amount": 500 } },
                  { "action": "screenshot" },
                  { "action": "list_tabs" },
                  { "action": "open_tab", "params": { "url": "https://gmail.com" } },
                  { "action": "switch_tab", "params": { "tab_id": 2 } },
                  { "action": "close_tab", "params": { "tab_id": 3 } },
                  { "action": "wait", "params": { "selector": ".results", "timeout": 10000 } },
                  { "action": "read_page", "params": { "limit": 20 } }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CommandResult" }
              }
            }
          },
          "400": {
            "description": "Missing action",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Browser not connected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Server fault, not the request: the body carries a ref that matches the server's log line",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key passed as Bearer token. Create one from the dashboard after signing in."
      }
    },
    "parameters": {
      "browserId": {
        "name": "browserId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Browser session ID (from GET /browsers)"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error", "code"],
        "properties": {
          "error": { "type": "string", "description": "What happened, why, and the next step. Stands alone: clients show it as is." },
          "code": { "type": "string", "description": "A stable reason to branch on: invalid_request, not_found, internal_error, command_outcome_unknown, action_unknown, action_unsupported, action_forbidden, tab_unprotected, endpoint_in_use, provider_failed, quota_exceeded, llm_unconfigured, and others named per route." },
          "field": { "type": "string", "description": "For invalid_request: the field that was wrong." },
          "ref": { "type": "string", "description": "For internal_error: eight hex characters that also sit in the server's log line for this failure." },
          "browserId": { "type": "string", "description": "For endpoint_in_use: the browser that holds the endpoint." },
          "action": { "type": "string", "description": "For action_unknown, action_unsupported and action_forbidden: the action as sent." },
          "clientType": { "type": "string", "description": "For action_unsupported: this browser's kind, oya or cdp." },
          "supportedOn": {
            "type": "array",
            "items": { "type": "string" },
            "description": "For action_unsupported: the kinds of browser that do the action."
          }
        }
      },
...
      "CommandResult": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "description": "Whether the command succeeded" },
          "code": { "type": "string", "description": "For ok:false, a stable reason: command_failed for a refusal the page answered, or one named per action." },
          "data": {
            "type": "object",
            "description": "Action-specific result data",
            "properties": {
              "url": { "type": "string" },
              "title": { "type": "string" },
              "screenshot": { "type": "string", "description": "Base64 PNG data URL (for screenshot action)" },
              "format": { "type": "string", "enum": ["markdown", "toon", "jsonl"], "description": "The format page is written in (for analyze action)" },
              "page": { "type": "string", "description": "The page, written in format (for analyze action)" },
              "markdown": { "type": "string", "description": "Page content as markdown, when format is markdown (for analyze action)" },
              "blocks": { "type": "array", "description": "The page as data: one block per heading, paragraph, list item, table row, image and element, in reading order (for analyze action)", "items": { "type": "object" } },
              "elements": {
                "type": "array",
                "description": "Interactive elements (for analyze/read_page actions)",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "type": { "type": "string" },
                    "text": { "type": "string" },
                    "href": { "type": "string" },
                    "visible": { "type": "boolean" }
                  }
                }
              },
              "tabs": {
                "type": "array",
                "description": "Open tabs (for list_tabs action)",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "title": { "type": "string" },
                    "url": { "type": "string" },
                    "active": { "type": "boolean" }
                  }
                }
              },
              "tab_id": { "type": "integer", "description": "New tab ID (for open_tab action)" },
              "selector": { "type": "string" },
              "direction": { "type": "string" },
              "amount": { "type": "number" },
              "key": { "type": "string" },
              "found": { "type": "boolean" },
              "closed": { "type": "boolean" },
              "needs_capture": { "type": "boolean" },
              "scroll": {
                "type": "object",
                "properties": {
                  "top": { "type": "number" },
                  "height": { "type": "number" },
                  "scrollHeight": { "type": "number" }
                }
              },
              "viewport": {
                "type": "object",
                "properties": {
                  "width": { "type": "number" },
                  "height": { "type": "number" }
                }
              },
              "truncated": { "type": "boolean" }
            }
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Error message if ok is false"
          }
        }
      }
    }
  }
}